+++ src/Component/AI/Backgammon/Engine.php
@@ -56,23 +56,25 @@ abstract class Engine
         }
         
         if ( $bestMoveSequence == null ) {
+            $this->logger->debug( $allSequences->toArray(), 'EngineAllSequences.txt' );
             return new ArrayCollection();
         }
         
+        //$this->logger->log( 'BestMoves Before Filter: ' . print_r( $bestMoveSequence->toArray(), true ), 'EnginMoves' );
         $bestMoveSequence   = $bestMoveSequence->filter(
             function( $entry ) {
-                return $entry != null;
+                return $entry != null && ! $entry->isNull();
             }
         );
         
         if ( $myColor == PlayerColor::Black ) {
-            $this->getMovesOrderedByFromBlackNumber( $bestMoveSequence, AbstractGameManager::COLLECTION_ORDER_ASC );
+            return $this->getMovesOrderedByFromBlackNumber( $bestMoveSequence, AbstractGameManager::COLLECTION_ORDER_ASC );
         }
         
         try {
             return $this->getMovesOrderedByFromWhiteNumber( $bestMoveSequence, AbstractGameManager::COLLECTION_ORDER_ASC );
         } catch ( \Exception $e ) {
-            $this->logger->log( 'Engin GetBestMoves: ' . print_r( $bestMoveSequence->toArray(), true ), 'EngineGenerateMoves' );
+            $this->logger->log( print_r( $bestMoveSequence->toArray(), true ), 'EngineGenerateMoves' );
             throw $e;
         }
     }
@@ -85,7 +87,10 @@ abstract class Engine
         foreach ( $game->Roll as $roll ) {
             $moves[]    = new Move();
         }
+        $sequences[]    = $moves;
+        
         $this->_GenerateMovesSequence( $sequences, $moves, 0, $game );
+        //$this->logger->log( 'Engin GetBestMoves: ' . print_r( $sequences->toArray(), true ), 'EngineGenerateMoves' );
         
         // Special case. Sometimes the first dice is blocked, but can be moved after next dice
         if ( $sequences->count() == 1 ) {
@@ -95,6 +100,7 @@ abstract class Engine
                 }
             );
             if ( $blockedMoves->count() ) {
+                $this->logger->log( 'Has Blocked Moves. Count: ' . $blockedMoves->count(), 'EngineGenerateMoves' );
                 $temp = $game->Roll[0];
                 $game->Roll[0] = $game->Roll[1];
                 $game->Roll[1] = $temp;
@@ -149,7 +155,7 @@ abstract class Engine
         try {
             $moves = new ArrayCollection();
             for ( $i = 0; $i < count( $sequence ); $i++ ) {
-                if ( $sequence[$i] != null ) {
+                if ( ! $sequence[$i]->isNull() ) {
                     $move   = new Move();
                     $move->From = $game->Points[$sequence[$i]->From->BlackNumber];
                     $move->To = $game->Points[$sequence[$i]->To->BlackNumber];
@@ -164,14 +170,14 @@ abstract class Engine
             throw $e;
         }
         
-        return $moves;
+        return $moves->count() ? $moves : $sequence;
     }
 
     protected function DoSequence( Collection $sequence, Game $game ): Collection
     {
         $hits = new ArrayCollection();
         foreach ( $sequence as $move ) {
-            if ( $move == null ) {
+            if ( $move == null || $move->isNull() ) {
                 continue;
             }
             $hit = $game->MakeMove( $move );
@@ -187,7 +193,7 @@ abstract class Engine
         $game->SwitchPlayer();
 
         for ( $i = $sequence->count() - 1; $i >= 0; $i-- ) {
-            if ( $sequence[$i] != null ) {
+            if ( $sequence[$i] != null && $hits->count() ) {
                 $lastHit    = $hits->last();
                 $hits->removeElement( $lastHit );