+++ themes/GamePlatform_VelzonSaas/assets/js/games/application/components/game-boards/backgammon-container/backgammon-container.component.ts
@@ -9,8 +9,10 @@ import {
     ElementRef,
     ChangeDetectorRef,
     Input,
+    Output,
     ViewChild,
-    HostListener
+    HostListener,
+    EventEmitter
 } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Observable, Subscription, map } from 'rxjs';
@@ -80,6 +82,8 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
     @Input() isLoggedIn: boolean            = false;
     @Input() introPlaying: boolean          = false;
     @Input() hasPlayer: boolean             = false;
+    
+    @Output() lobbyButtonsVisibleChanged    = new EventEmitter<boolean>();
     
     @ViewChild( 'dices' ) dices: ElementRef | undefined;
     @ViewChild( 'backgammonBoardButtons' ) backgammonBoardButtons: ElementRef | undefined;
@@ -177,6 +181,7 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
         this.gameString$ = this.appStateService.gameString.observe();
         
         this.user$.subscribe( ( user ) => {
+            //alert( 'User: ' + user );
             if ( user ) this.introMuted = user.muteIntro;
         });
         
@@ -186,6 +191,9 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
         }
         
         //console.log( 'Game Settings: ', window.gamePlatformSettings );
+        const inviteId = window.gamePlatformSettings.queryParams.inviteId;
+        //alert( '[In Game Container]Invite ID: ' + inviteId );
+        
         const gameId = window.gamePlatformSettings.queryParams.gameId;
         const playAi = window.gamePlatformSettings.queryParams.playAi;
         const forGold = window.gamePlatformSettings.queryParams.forGold;
@@ -203,8 +211,12 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
             setTimeout( () => {
                 this.tutorialService.start();
             }, 1 );
-        } else if ( ! this.editing ) {
+        } else if ( ! this.editing && ! inviteId ) {
             this.wsService.connect( gameId, playAi, forGold );
+            if ( gameId ) {
+                //alert( 'Game ID: ' + gameId );
+                this.playGame( gameId );
+            }
         }
         
         if ( this.editing ) {
@@ -306,6 +318,9 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
                 case 'isLoggedIn':
                     this.isLoggedIn = changedProp.currentValue;
                     break;
+                case 'lobbyButtonsVisible':
+                    this.lobbyButtonsVisible = changedProp.currentValue;
+                    break;
                 case 'introPlaying':
                     this.introPlaying = changedProp.currentValue;
                     break;
@@ -386,14 +401,12 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
             return;
         }
         
-        //alert( dto?.playState );
         if ( ! this.started && dto ) {
             if ( dto.playState === GameState.playing ) {
-                //alert( dto.playState );
                 clearTimeout( this.startedHandle );
                 this.started = true;
                 this.playAiQuestion = false;
-                this.lobbyButtonsVisible = false;
+                this.lobbyButtonsVisibleChanged.emit( false );
             }
             
             if ( dto.isGoldGame ) this.sound.playCoin();
@@ -491,9 +504,13 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
     @HostListener( 'window:resize', ['$event'] )
     onResize(): void
     {
-        // const _innerWidth   = window.innerWidth;
+        //const _innerWidth   = window.innerWidth;
         const _innerWidth   = $( '#GameBoardContainer' ).width();
-        // alert( _innerWidth );
+        //const _innerHeight   = window.innerHeight;
+        const _innerHeight   = $( '#GameBoardContainer' ).height();
+        
+        //console.log( 'Window innerHeight', window.innerHeight );
+        //console.log( 'Container innerHeight', $( '#GameBoardContainer' ).height() );
         
         this.width = Math.min( _innerWidth, 1024 );
         const span = this.messages?.nativeElement as Element;
@@ -504,12 +521,18 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
         this.messageCenter = this.width / 2 - spanWidth / 2;
         // alert( this.messageCenter );
         
-        this.height = Math.min( window.innerHeight - 40, this.width * 0.62 );
+        this.height = Math.min( _innerHeight - 40, this.width * 0.6 );
         
         const buttons = this.backgammonBoardButtons?.nativeElement as HTMLElement;
         const btnsOffset = 5; //Cheating. Could not get the height.
         if ( buttons ) {
-            buttons.style.top = `${this.height / 2 - btnsOffset}px`;
+            let buttonsTop  = this.height / 2 - btnsOffset;
+            // My Workaround
+            if ( $( 'canvas.game-board' ).hasClass( 'flipped' ) || $( 'canvas.game-board' ).hasClass( 'rotated' ) ) {
+                buttonsTop  = buttonsTop - 10;
+            }
+            
+            buttons.style.top = `${buttonsTop}px`;
             buttons.style.right = `${this.width * 0.11}px`;
         }
         
@@ -524,7 +547,13 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
                 dices.style.left = '';
                 //alert( dices.style.right );
             }
-            dices.style.top = `${this.height / 2 - btnsOffset}px`;
+            
+            let dicesTop  = this.height / 2 - btnsOffset - 10;
+            // My Workaround
+            if ( $( 'canvas.game-board' ).hasClass( 'flipped' ) ) {
+                dicesTop  = dicesTop - 15;
+            }
+            dices.style.top = `${dicesTop}px`;
         }
     }
     
@@ -619,7 +648,7 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
         
         this.gamePlayService.exitBoardGame();
         this.playAiQuestion = false;
-        this.lobbyButtonsVisible = true;
+        this.lobbyButtonsVisibleChanged.emit( true );
     }
     
     requestDoubling(): void
@@ -694,10 +723,26 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
             modalRef.dismiss();
         });
         
-        modalRef.componentInstance.onPlayGame.subscribe( () => {
+        modalRef.componentInstance.onPlayGame.subscribe( ( gameId: string ) => {
             modalRef.close();
-            this.playGame();
+            
+            this.wsService.connect( gameId, this.playAiFlag, this.forGoldFlag );
+            this.playGame( gameId );
         });
+    }
+    
+    acceptInvite( inviteId: string ): void
+    {
+        this.wsService.acceptInvite( inviteId );
+        
+        this.wsService.resetGame();
+        this.wsService.connect( inviteId, this.playAiFlag, this.forGoldFlag );
+        this.waitForOpponent();
+    }
+    
+    cancelInvite(): void
+    {
+        this.exitGame();
     }
     
     login(): void
@@ -750,16 +795,20 @@ export class BackgammonContainerComponent implements OnDestroy, AfterViewInit, O
         this.authService.toggleIntro();
     }
     
-    playGame(): void
+    playGame( gameId: string ): void
     {
         const game      = this.appStateService.game.getValue();
         const myColor   = this.appStateService.myColor.getValue();
         //console.log( 'GameDto Object: ', game );
         
-        this.gamePlayService.startBoardGame( 'normal' );
+        if ( ! gameId.length ) {
+            this.gamePlayService.startBoardGame( 'normal' );
+        }
+        
         //this.wsService.connect( '', this.playAiFlag, this.forGoldFlag );
         this.wsService.startGamePlay( game, myColor, this.playAiFlag, this.forGoldFlag );
         
+        this.lobbyButtonsVisibleChanged.emit( false );
         this.waitForOpponent();
         window.dispatchEvent( new Event( 'resize' ) );